Made ticket-single clean up its state when done
The ticket-single view was not cleaning up its state once it
exited. This commit adds that clean-up. The lack of clean up
was causing the ticket-edit component to render before the ticket
state was properly set. The edit field was not updating when the
new ticketSingle state was updated. This closes issue #66.